% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShark.R
\name{addDyntaxa}
\alias{addDyntaxa}
\title{Add taxonomic ranks to datasets from dyntaxa.se}
\usage{
addDyntaxa(data)
}
\arguments{
\item{data}{(Required) A string representing the type of data to be downloaded. (Only "Zooplankton" and "Phytoplankton" tested).}

\item{possibly}{(Optional, default=TRUE) Logical. If TRUE: datasets that does not fit the format (cannot be read as TSV) will not be downloaded. If FALSE: the function will stop if an errournous dataset is encountered.}
}
\value{
A dataframe.
}
\description{
Add taxonomic ranks to datasets from dyntaxa.se
}
\examples{

data <- getSHARK("Phytoplankton") \%>\%
   addDyntaxa()

}
