% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShark.R
\name{annotateSHARK}
\alias{annotateSHARK}
\title{Annotate the downloaded dataset}
\usage{
annotateSHARK(data)
}
\arguments{
\item{data}{(Required) The dataset.}

\item{possibly}{(Optional, default=TRUE) Logical. If TRUE: datasets that does not fit the format (cannot be read as TSV) will not be downloaded. If FALSE: the function will stop if an errournous dataset is encountered.}
}
\value{
Annotated dataframe.
}
\description{
Annotate the downloaded dataset
}
\examples{
data(dyntaxa)
data <- getSHARK("Phytoplankton") \%>\%
addDyntaxa(dyntaxa) \%>\%
annotateSHARK()

}
