% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShark.R
\name{dailyInterpretation}
\alias{dailyInterpretation}
\title{Estimation of daily mean abundances from long term monitoring data}
\usage{
dailyInterpretation(data, taxa = "Genus")
}
\arguments{
\item{data}{(Required) A dataframe (of type Zooplankton or Phytoplankton?)}

\item{taxa}{(Optional, default=Genus") A character string refering to the taxa-specific column in the dataframe}
}
\value{
Dataframe, with year summarized and daily abundances interpreted.
}
\description{
Uses linear interpretation to calculate hypothetical abundances for all days during montoring perod.
}
\examples{
data <- getSHARK("Phytoplankton") \%>\%
   addDyntaxa() \%>\%
   annotateSHARK() \%>\%

 filter(Station == "BY31 LANDSORTSDJ",
        Year > 2006,
        unit == 	"mm3/l",
        Phylum \%in\% c(
                      "Miozoa",
                      "Bacillariophyta",
                      "Ciliophora",
                      "Haptophyta",
                      "Chlorophyta",
                      "Cyanobacteria"
  )) \%>\%

 group_by(SDATE, Phylum, Depth) \%>\%
 summarize(Value=sum(Value)) \%>\%

 group_by(SDATE, Phylum) \%>\%
 summarize(Value=mean(Value)) \%>\%

 dailyInterpretation(taxa="Phylum")

}
