% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreSHARK.R
\name{exploreSHARK}
\alias{exploreSHARK}
\title{Launch interactive SHARK data explorer}
\usage{
exploreSHARK(download = FALSE)
}
\arguments{
\item{download}{(Optional, default=FALSE) If TRUE, zooplankton and phytoplankton data will be retrieved from SHARKdata, and saved to shaRk R package directory. This only has to be done the first time the function is run for each version installation or update, or to get updated versions from SHARKdata.}
}
\description{
Launch interactive SHARK data explorer
}
\examples{
exploreSHARK()

exploreSHARK(TRUE)

}
