% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShark.R
\name{getSHARK}
\alias{getSHARK}
\title{Downloads current monitoring data from SHARKdata}
\usage{
getSHARK(datatype, possibly = TRUE)
}
\arguments{
\item{datatype}{(Required) A string representing the type of data to be downloaded. (Only "Zooplankton" and "Phytoplankton" tested).}

\item{possibly}{(Optional, default=TRUE) Logical. If TRUE: datasets that does not fit the format (cannot be read as TSV) will not be downloaded. If FALSE: the function will stop if an errournous dataset is encountered.}
}
\value{
A dataframe.
}
\description{
This function downloads all current data at sharkdata.se by data type.
}
\examples{
data <- getSHARK("Phytoplankton")

}
